requirements.txt
einops==0.6.1
hydra-core==1.3.2
matplotlib==3.7.1
omegaconf==2.3.0
pandas==2.0.3
Pillow==9.4.0
pytorch-ignite==0.4.12
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
torch==2.0.1
torchmetrics==1.0.3
torchvision==0.15.2
tqdm==4.66.1
wandb==0.15.8
